% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_version.R
\name{rl_api_version}
\alias{rl_api_version}
\title{Get the Red List API version}
\usage{
rl_api_version(key = NULL, ...)
}
\arguments{
\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
API version as character string
}
\description{
Returns the current version number of the IUCN Red List of Threatened Species
API
}
\examples{
\dontrun{
rl_api_version()
}
}
\seealso{
Red List information and statistics:
\code{\link{rl_citation}()},
\code{\link{rl_sp_count}()},
\code{\link{rl_version}()}
}
\concept{stats}
