% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_categories.R
\name{rl_categories}
\alias{rl_categories}
\alias{rl_categories_}
\title{Red List category assessment summary}
\usage{
rl_categories(
  code = NULL,
  key = NULL,
  parse = TRUE,
  all = TRUE,
  page = 1,
  quiet = FALSE,
  ...
)

rl_categories_(
  code = NULL,
  key = NULL,
  all = TRUE,
  page = 1,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{code}{(character) The code of the Red List category to look up. If not
supplied, a list of all categories will be returned.}

\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{all}{(logical) Whether to retrieve all results at once or not. If
\code{TRUE} we do the paging internally for you and bind all of the results
together. If \code{FALSE}, only a single page of results will be retrieved.}

\item{page}{(integer/numeric) Page to get if \code{all} is \code{FALSE}. Default: 1.
Each page returns up to 100 records. Paging is required because it's too
much burden on a server to just "get all the data" in one request.}

\item{quiet}{(logical) Whether to suppress progress for multi-page downloads
or not. Default: \code{FALSE} (that is, give progress). Ignored if \code{all = FALSE}.}

\item{...}{Includes the following filters that are supported by the API
endpoint (refer to the \href{https://api.iucnredlist.org/api-docs/index.html}{API docs} for more
information):
\itemize{
\item \code{year_published}: (integer) Set this to return only
assessments from a given year.
\item \code{latest}: (logical) Set this to \code{TRUE} to return only the
latest assessment for each taxon.
\item \code{scope_code}: (integer) Set this to return only assessments
from a particular \link[=rl_scopes]{scope} (e.g., \code{1} for Global, \code{2} for
Europe). This is similar to the \code{region} argument of the old Red List
API and old versions of rredlist.
}
Also supports any \link[curl:curl_options]{curl options} passed to the GET
request via \code{\link[crul]{HttpClient}}.}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
Return a list of the latest assessments for a given category (e.g., Not
Evaluated, Data Deficient, Least Concern, Near Threatened, Vulnerable,
Endangered, Critically Endangered, Extinct in the Wild and Extinct). This
endpoint returns categories for the current IUCN Red List Categories and
Criteria (v3.1) as well as older versions (i.e., v2.3). Note that a code may
not be unique across Categories and Criteria versions. Therefore, codes like
"EX" will return assessments for EX v3.1 and EX v2.3.
}
\examples{
\dontrun{
# Get all Red List categories
rl_categories()
# Get assessments for the Vulnerable category
rl_categories("VU")
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
