% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_taxa.R
\name{rl_family}
\alias{rl_family}
\alias{rl_family_}
\title{Family assessment summary}
\usage{
rl_family(
  family = NULL,
  key = NULL,
  parse = TRUE,
  all = TRUE,
  page = 1,
  quiet = FALSE,
  ...
)

rl_family_(family = NULL, key = NULL, all = TRUE, page = 1, quiet = FALSE, ...)
}
\arguments{
\item{family}{(character) The name of the family to look up. If not supplied,
a list of all family names will be returned.}

\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{all}{(logical) Whether to retrieve all results at once or not. If
\code{TRUE} we do the paging internally for you and bind all of the results
together. If \code{FALSE}, only a single page of results will be retrieved.}

\item{page}{(integer/numeric) Page to get if \code{all} is \code{FALSE}. Default: 1.
Each page returns up to 100 records. Paging is required because it's too
much burden on a server to just "get all the data" in one request.}

\item{quiet}{(logical) Whether to suppress progress for multi-page downloads
or not. Default: \code{FALSE} (that is, give progress). Ignored if \code{all = FALSE}.}

\item{...}{Includes the following filters that are supported by the API
endpoint (refer to the \href{https://api.iucnredlist.org/api-docs/index.html}{API docs} for more
information):
\itemize{
\item \code{year_published}: (integer) Set this to return only
assessments from a given year.
\item \code{latest}: (logical) Set this to \code{TRUE} to return only the
latest assessment for each taxon.
\item \code{scope_code}: (integer) Set this to return only assessments
from a particular \link[=rl_scopes]{scope} (e.g., \code{1} for Global, \code{2} for
Europe).
}
Also supports any \link[curl:curl_options]{curl options} passed to the GET
request via \code{\link[crul]{HttpClient}}.}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
Get an assessment summary for a particular family
}
\examples{
\dontrun{
# Get assessment summary for family
ex1 <- rl_family(family = "Hominidae")
nrow(ex1$assessments)
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
\seealso{
Assessments by taxa:
\code{\link{rl_class}()},
\code{\link{rl_extinct}()},
\code{\link{rl_extinct_wild}()},
\code{\link{rl_kingdom}()},
\code{\link{rl_order}()},
\code{\link{rl_phylum}()},
\code{\link{rl_sis}()},
\code{\link{rl_species}()}
}
\concept{taxa}
