% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_categories.R
\name{rl_green}
\alias{rl_green}
\alias{rl_green_}
\title{Green Status assessment summary}
\usage{
rl_green(key = NULL, parse = TRUE, ...)

rl_green_(key = NULL, ...)
}
\arguments{
\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}. Also
includes the following arguments that are supported by most of the API
endpoints which return lists of assessments (refer to the \href{https://api.iucnredlist.org/api-docs/index.html}{API docs} for more
information):
\itemize{
\item \code{year_published}: (integer) Set this to return only
assessments from a given year.
\item \code{latest}: (logical) Set this to \code{TRUE} to return only the
latest assessment for each taxon.
\item \code{scope_code}: (integer) Set this to return only assessments
from a particular scope (e.g., Global, Europe).
}}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
List all Green Status assessments.
}
\examples{
\dontrun{
# Get list of Green Status assessments
rl_green()
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
