% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_taxa.R
\name{rl_kingdom}
\alias{rl_kingdom}
\alias{rl_kingdom_}
\title{Kingdom assessment summary}
\usage{
rl_kingdom(
  kingdom = NULL,
  key = NULL,
  parse = TRUE,
  all = TRUE,
  page = 1,
  quiet = FALSE,
  ...
)

rl_kingdom_(
  kingdom = NULL,
  key = NULL,
  all = TRUE,
  page = 1,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{kingdom}{(character) The name of the kingdom to look up. If not
supplied, a list of all kingdom names will be returned.}

\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{all}{(logical) Whether to retrieve all results at once or not. If
\code{TRUE} we do the paging internally for you and bind all of the results
together. If \code{FALSE}, only a single page of results will be retrieved.}

\item{page}{(integer/numeric) Page to get if \code{all} is \code{FALSE}. Default: 1.
Each page returns up to 100 records. Paging is required because it's too
much burden on a server to just "get all the data" in one request.}

\item{quiet}{(logical) Whether to give progress for download or not. Default:
\code{FALSE} (that is, give progress). Ignored if \code{all = FALSE}.}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}. Also
includes the following arguments that are supported by most of the API
endpoints which return lists of assessments (refer to the \href{https://api.iucnredlist.org/api-docs/index.html}{API docs} for more
information):
\itemize{
\item \code{year_published}: (integer) Set this to return only
assessments from a given year.
\item \code{latest}: (logical) Set this to \code{TRUE} to return only the
latest assessment for each taxon.
\item \code{scope_code}: (integer) Set this to return only assessments
from a particular scope (e.g., Global, Europe).
}}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
Get an assessment summary for a particular kingdom
}
\examples{
\dontrun{
# Get assessment summary for kingdom
ex1 <- rl_kingdom(kingdom = "Fungi")
nrow(ex1$assessments)
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
\seealso{
Taxa:
\code{\link{rl_class}()},
\code{\link{rl_extinct}()},
\code{\link{rl_extinct_wild}()},
\code{\link{rl_family}()},
\code{\link{rl_order}()},
\code{\link{rl_phylum}()},
\code{\link{rl_sis}()},
\code{\link{rl_species}()}
}
\concept{taxa}
