% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_taxa.R
\name{rl_phylum}
\alias{rl_phylum}
\alias{rl_phylum_}
\title{Phylum assessment summary}
\usage{
rl_phylum(
  phylum = NULL,
  key = NULL,
  parse = TRUE,
  all = TRUE,
  page = 1,
  quiet = FALSE,
  ...
)

rl_phylum_(phylum = NULL, key = NULL, all = TRUE, page = 1, quiet = FALSE, ...)
}
\arguments{
\item{phylum}{(character) The name of the phylum to look up. If not supplied,
a list of all phylum names will be returned.}

\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{all}{(logical) Whether to retrieve all results at once or not. If
\code{TRUE} we do the paging internally for you and bind all of the results
together. If \code{FALSE}, only a single page of results will be retrieved.}

\item{page}{(integer/numeric) Page to get if \code{all} is \code{FALSE}. Default: 1.
Each page returns up to 100 records. Paging is required because it's too
much burden on a server to just "get all the data" in one request.}

\item{quiet}{(logical) Whether to give progress for download or not. Default:
\code{FALSE} (that is, give progress). Ignored if \code{all = FALSE}.}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
Get an assessment summary for a particular phylum
}
\examples{
\dontrun{
# Get assessment summary for phylum
ex1 <- rl_phylum(phylum = "Annelida`")
nrow(ex1$assessments)
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
\seealso{
Taxa:
\code{\link{rl_class}()},
\code{\link{rl_extinct}()},
\code{\link{rl_extinct_wild}()},
\code{\link{rl_family}()},
\code{\link{rl_kingdom}()},
\code{\link{rl_order}()},
\code{\link{rl_sis}()},
\code{\link{rl_species}()}
}
\concept{taxa}
