% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rl_regions}
\alias{rl_regions}
\alias{rl_regions_}
\title{Get regions}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is fully deprecated in favor of
\link[=rl_scopes]{rl_scopes(code = NULL)} as of \strong{rredlist} version 1.0.0.
}
\keyword{internal}
