% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_conservation.R
\name{rl_research}
\alias{rl_research}
\alias{rl_research_}
\title{Research needed assessment summary}
\usage{
rl_research(
  code = NULL,
  key = NULL,
  parse = TRUE,
  all = TRUE,
  page = 1,
  quiet = FALSE,
  ...
)

rl_research_(code = NULL, key = NULL, all = TRUE, page = 1, quiet = FALSE, ...)
}
\arguments{
\item{code}{(character) The code of the research type to look up. If not
supplied, a list of all research types will be returned.}

\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{all}{(logical) Whether to retrieve all results at once or not. If
\code{TRUE} we do the paging internally for you and bind all of the results
together. If \code{FALSE}, only a single page of results will be retrieved.}

\item{page}{(integer/numeric) Page to get if \code{all} is \code{FALSE}. Default: 1.
Each page returns up to 100 records. Paging is required because it's too
much burden on a server to just "get all the data" in one request.}

\item{quiet}{(logical) Whether to give progress for download or not. Default:
\code{FALSE} (that is, give progress). Ignored if \code{all = FALSE}.}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}. Also
includes the following arguments that are supported by most of the API
endpoints which return lists of assessments (refer to the \href{https://api.iucnredlist.org/api-docs/index.html}{API docs} for more
information):
\itemize{
\item \code{year_published}: (integer) Set this to return only
assessments from a given year.
\item \code{latest}: (logical) Set this to \code{TRUE} to return only the
latest assessment for each taxon.
\item \code{scope_code}: (integer) Set this to return only assessments
from a particular scope (e.g., Global, Europe).
}}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
Return the latest assessments with a given research needed (e.g., Taxonomy or
Population trends). These research codes correspond to the IUCN Red List
Research Needed Classification Scheme (v1.0).
}
\examples{
\dontrun{
# Get list of all research types
rl_research()
# Get assessments with Taxonomy research needed
rl_research("1_1")
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
\seealso{
Conservation actions and research needed:
\code{\link{rl_actions}()}
}
\concept{conservation}
