% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_taxa.R
\name{rl_sis}
\alias{rl_sis}
\alias{rl_sis_}
\title{SIS ID assessment summary}
\usage{
rl_sis(id, key = NULL, parse = TRUE, ...)

rl_sis_(id, key = NULL, ...)
}
\arguments{
\item{id}{(integer) The SIS ID of the taxonomic entity to look up.}

\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}. Also
includes the following arguments that are supported by most of the API
endpoints which return lists of assessments (refer to the \href{https://api.iucnredlist.org/api-docs/index.html}{API docs} for more
information):
\itemize{
\item \code{year_published}: (integer) Set this to return only
assessments from a given year.
\item \code{latest}: (logical) Set this to \code{TRUE} to return only the
latest assessment for each taxon.
\item \code{scope_code}: (integer) Set this to return only assessments
from a particular scope (e.g., Global, Europe).
}}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
Get an assessment summary for a particular taxonomic entity based on its ID
number from the \href{https://www.iucnredlist.org/assessment/sis}{IUCN Species Information Service (SIS)}.
}
\examples{
\dontrun{
# Get assessment summary for species
ex1 <- rl_sis(id = 9404)
nrow(ex1$assessments)
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
\seealso{
Taxa:
\code{\link{rl_class}()},
\code{\link{rl_extinct}()},
\code{\link{rl_extinct_wild}()},
\code{\link{rl_family}()},
\code{\link{rl_kingdom}()},
\code{\link{rl_order}()},
\code{\link{rl_phylum}()},
\code{\link{rl_species}()}
}
\concept{taxa}
