% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rl_sp_country}
\alias{rl_sp_country}
\alias{rl_sp_country_}
\title{Get species by country}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is fully deprecated in favor of \code{\link[=rl_countries]{rl_countries()}} as of
\strong{rredlist} version 1.0.0.
}
\keyword{internal}
