% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_taxa.R
\name{rl_species}
\alias{rl_species}
\alias{rl_species_}
\title{Species assessment summary}
\usage{
rl_species(
  genus,
  species,
  infra = NULL,
  subpopulation = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

rl_species_(
  genus,
  species,
  infra = NULL,
  subpopulation = NULL,
  key = NULL,
  ...
)
}
\arguments{
\item{genus}{(character) The genus name of the species to look up.}

\item{species}{(character) The species epithet of the species to look up.}

\item{infra}{(character) An optional name of the subspecies or variety to
look up.}

\item{subpopulation}{(character) An optional name of the geographically
separate subpopulation to look up.}

\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{\link[curl:curl_options]{Curl options} passed to the GET request via
\code{\link[crul]{HttpClient}}.}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
Get an assessment summary for a particular species (i.e., Latin binomial) or
subspecies/variety/subpopulation (i.e., Latin trinomial).
}
\details{
Geographically separate subpopulations of a species are defined as
those populations that are so isolated from others of the same species that
it is considered extremely unlikely that there is any genetic interchange.
In general, listings of such subpopulations are restricted to those that
have been isolated for a long period of time.

Assessments of subspecies, varieties, and geographically separate
subpopulations must adhere to the same standards as for species
assessments. However, these assessments are only included provided there is
a global assessment of the species as a whole.

Infraspecific ranks such as formas, subvarieties, cultivars, etc are not
included in the Red List.
}
\examples{
\dontrun{
# Get assessment summary for species
ex1 <- rl_species(genus = "Fratercula", species = "arctica")
nrow(ex1$assessments)

# Get assessment summary for subspecies
ex2 <- rl_species(genus = "Gorilla", species = "gorilla",
                  infra = "gorilla")
nrow(ex2$assessments)
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
\seealso{
Assessments by taxa:
\code{\link{rl_class}()},
\code{\link{rl_family}()},
\code{\link{rl_kingdom}()},
\code{\link{rl_order}()},
\code{\link{rl_phylum}()},
\code{\link{rl_sis}()},
\code{\link{rl_sis_latest}()},
\code{\link{rl_species_latest}()}
}
\concept{taxa}
