% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rl_synonyms}
\alias{rl_synonyms}
\alias{rl_synonyms_}
\title{Get species synonym information by taxonomic name}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is fully deprecated in favor of
\code{\link[=rl_species_latest]{rl_species_latest()}}\verb{$taxon$synonyms} as of \strong{rredlist} version 1.0.0.
}
\keyword{internal}
