% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_version.R
\name{rl_version}
\alias{rl_version}
\title{Get the Red List version}
\usage{
rl_version(key = NULL, ...)
}
\arguments{
\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
Red List version as character string
}
\description{
Returns the current version number of the IUCN Red List of Threatened Species
}
\examples{
\dontrun{
rl_version()
}
}
\seealso{
Red List information and statistics:
\code{\link{rl_api_version}()},
\code{\link{rl_citation}()},
\code{\link{rl_sp_count}()}
}
\concept{stats}
