% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_citation.R
\name{rl_citation}
\alias{rl_citation}
\title{Get the citation Red List API version}
\usage{
rl_citation(key = NULL, ...)
}
\arguments{
\item{key}{A IUCN API token}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\value{
API citation as character string
}
\description{
Get the citation Red List API version
}
\examples{
\dontrun{
rl_citation()
}
}

