% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_citation.R
\name{rl_citation}
\alias{rl_citation}
\title{Get the citation Red List API version}
\usage{
rl_citation(...)
}
\arguments{
\item{...}{Curl options passed to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
API citation as character string
}
\description{
Get the citation Red List API version
}
\examples{
\dontrun{
rl_citation()
}
}
