% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_comp_groups.R
\name{rl_comp_groups}
\alias{rl_comp_groups}
\alias{rl_comp_groups_}
\title{Information about comprehensive groups}
\usage{
rl_comp_groups(group = NULL, key = NULL, parse = TRUE, ...)

rl_comp_groups_(group = NULL, key = NULL, ...)
}
\arguments{
\item{group}{(character) A comprehensive group name.
Call \code{rl_comp_groups()} without passing this parameter
to get the list of comprehensive groups}

\item{key}{A IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
A list, with the data in the \code{result} slot, unless using
a function with a trailing underscore, in which case json as character
string is returned.
}
\description{
Information about comprehensive groups
}
\examples{
\dontrun{
rl_comp_groups()
rl_comp_groups('mammals')
rl_comp_groups('groupers')

rl_comp_groups_()
rl_comp_groups_('groupers')
}
}
\references{
API docs at https://apiv3.iucnredlist.org/api/v3/docs
}
