% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_countries.R
\name{rl_countries}
\alias{rl_countries}
\alias{rl_countries_}
\title{Get countries}
\usage{
rl_countries(key = NULL, parse = TRUE, ...)

rl_countries_(key = NULL, ...)
}
\arguments{
\item{key}{A IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
A list, with the data in the \code{result} slot, unless using
a function with a trailing underscore, in which case json as character
string is returned.
}
\description{
Get countries
}
\examples{
\dontrun{
rl_countries()
rl_countries_()
}
}
\references{
API docs at \url{http://apiv3.iucnredlist.org/api/v3/docs}
}
