% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_categories.R
\name{rl_green}
\alias{rl_green}
\alias{rl_green_}
\title{Green Status assessment summary}
\usage{
rl_green(key = NULL, parse = TRUE, ...)

rl_green_(key = NULL, ...)
}
\arguments{
\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{\link[curl:curl_options]{Curl options} passed to the GET request via
\code{\link[crul]{HttpClient}}.}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
List all Green Status assessments.
}
\examples{
\dontrun{
# Get list of Green Status assessments
rl_green()
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
