% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rl_narrative}
\alias{rl_narrative}
\alias{rl_narrative_}
\title{Get species narrative information by taxon name, IUCN id, and region}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is fully deprecated in favor of
\code{\link[=rl_species_latest]{rl_species_latest()}}\verb{$documentation} as of \strong{rredlist} version 1.0.0.
}
\keyword{internal}
