% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_regions.R
\name{rl_regions}
\alias{rl_regions}
\alias{rl_regions_}
\title{Get regions}
\usage{
rl_regions(key = NULL, parse = TRUE, ...)

rl_regions_(key = NULL, ...)
}
\arguments{
\item{key}{A IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
A list, with the data in the \code{result} slot, unless using
a function with a trailing underscore, in which case json as character
string is returned.
}
\description{
Get regions
}
\examples{
\dontrun{
rl_regions()
rl_regions(parse = FALSE)
rl_regions_()
}
}
\references{
API docs at https://apiv3.iucnredlist.org/api/v3/docs
}
