% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_conservation.R
\name{rl_research}
\alias{rl_research}
\alias{rl_research_}
\title{Research needed assessment summary}
\usage{
rl_research(
  code = NULL,
  key = NULL,
  parse = TRUE,
  all = TRUE,
  page = 1,
  quiet = FALSE,
  ...
)

rl_research_(code = NULL, key = NULL, all = TRUE, page = 1, quiet = FALSE, ...)
}
\arguments{
\item{code}{(character) The code of the research type to look up. If not
supplied, a list of all research types will be returned.}

\item{key}{(character) An IUCN API token. See \code{\link[=rl_use_iucn]{rl_use_iucn()}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}.}

\item{all}{(logical) Whether to retrieve all results at once or not. If
\code{TRUE} we do the paging internally for you and bind all of the results
together. If \code{FALSE}, only a single page of results will be retrieved.}

\item{page}{(integer/numeric) Page to get if \code{all} is \code{FALSE}. Default: 1.
Each page returns up to 100 records. Paging is required because it's too
much burden on a server to just "get all the data" in one request.}

\item{quiet}{(logical) Whether to suppress progress for multi-page downloads
or not. Default: \code{FALSE} (that is, give progress). Ignored if \code{all = FALSE}.}

\item{...}{Includes the following filters that are supported by the API
endpoint (refer to the \href{https://api.iucnredlist.org/api-docs/index.html}{API docs} for more
information):
\itemize{
\item \code{year_published}: (integer) Set this to return only
assessments from a given year.
\item \code{latest}: (logical) Set this to \code{TRUE} to return only the
latest assessment for each taxon.
\item \code{scope_code}: (integer) Set this to return only assessments
from a particular \link[=rl_scopes]{scope} (e.g., \code{1} for Global, \code{2} for
Europe). This is similar to the \code{region} argument of the old Red List
API and old versions of rredlist.
}
Also includes the following arguments related to the wait time between
request retries if a "Too Many Requests" error is received from the API
(see \link[crul:HttpClient]{HttpClient()$retry()} for more details):
\itemize{
\item \code{pause_base}, \code{pause_cap}, and \code{pause_min}: basis, maximum, and minimum
for calculating wait time for retry
\item \code{times}: the maximum number of times to retry
\item \code{onwait}: a callback function if the request will be retried and a
wait time is being applied
}
Also supports any \link[curl:curl_options]{curl options} passed to the GET
request via \link[crul:HttpClient]{HttpClient}.}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
Return the latest assessments with a given research needed (e.g., Taxonomy or
Population trends). These research codes correspond to the IUCN Red List
Research Needed Classification Scheme (v1.0).
}
\examples{
\dontrun{
# Get list of all research types
rl_research()
# Get assessments with Taxonomy research needed
rl_research("1_1")
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
\seealso{
Assessments by conservation actions or research needed:
\code{\link{rl_actions}()}
}
\concept{conservation}
