% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rl_search}
\alias{rl_search}
\alias{rl_search_}
\title{Search by taxon name, IUCN id, and region}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is fully deprecated in favor of \code{\link[=rl_species]{rl_species()}}, \code{\link[=rl_sis]{rl_sis()}},
and \code{\link[=rl_scopes]{rl_scopes()}} as of \strong{rredlist} version 1.0.0.
}
\keyword{internal}
