% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_taxa.R
\name{rl_sis}
\alias{rl_sis}
\alias{rl_sis_}
\title{SIS ID assessment summary}
\usage{
rl_sis(id, key = NULL, parse = TRUE, ...)

rl_sis_(id, key = NULL, ...)
}
\arguments{
\item{id}{(integer) The SIS ID of the taxonomic entity to look up.}

\item{key}{(character) An IUCN API token. See \code{\link{rl_use_iucn}}.}

\item{parse}{(logical) Whether to parse the output to list (\code{FALSE}) or,
where possible, data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{\link[curl:curl_options]{Curl options} passed to the GET request via
\code{\link[crul]{HttpClient}}.}
}
\value{
A list unless using a function with a trailing underscore, in which
case json as character string is returned.
}
\description{
Get an assessment summary for a particular taxonomic entity based on its ID
number from the \href{https://www.iucnredlist.org/assessment/sis}{IUCN Species Information Service (SIS)}.
}
\examples{
\dontrun{
# Get assessment summary for species
ex1 <- rl_sis(id = 9404)
nrow(ex1$assessments)
}
}
\references{
API docs at \url{https://api.iucnredlist.org/}.
}
\seealso{
Assessments by taxa:
\code{\link{rl_class}()},
\code{\link{rl_family}()},
\code{\link{rl_kingdom}()},
\code{\link{rl_order}()},
\code{\link{rl_phylum}()},
\code{\link{rl_sis_latest}()},
\code{\link{rl_species}()},
\code{\link{rl_species_latest}()}
}
\concept{taxa}
