% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rl_sp}
\alias{rl_sp}
\alias{rl_sp_}
\title{Get species}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is fully deprecated in favor of \code{\link[=rl_kingdom]{rl_kingdom()}},
\code{\link[=rl_phylum]{rl_phylum()}}, \code{\link[=rl_class]{rl_class()}}, \code{\link[=rl_order]{rl_order()}}, and \code{\link[=rl_family]{rl_family()}} as of
\strong{rredlist} version 1.0.0.
}
\keyword{internal}
