% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_sp.R
\name{rl_sp}
\alias{rl_sp}
\alias{rl_sp_}
\title{Get species}
\usage{
rl_sp(page, key = NULL, parse = TRUE, ...)

rl_sp_(page, key = NULL, ...)
}
\arguments{
\item{page}{(interger) Page to get. Default: 1}

\item{key}{A IUCN API token}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\description{
Get species
}
\examples{
\dontrun{
rl_sp(page = 3)
}
}

