% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_sp_category.R
\name{rl_sp_category}
\alias{rl_sp_category}
\alias{rl_sp_category_}
\title{Get species by category}
\usage{
rl_sp_category(category, key = NULL, parse = TRUE, ...)

rl_sp_category_(category, key = NULL, parse = TRUE, ...)
}
\arguments{
\item{category}{(character) 2 letter category code}

\item{key}{A IUCN API token}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\description{
Get species by category
}
\examples{
\dontrun{
rl_sp_category('VU')
rl_sp_category('EN')
rl_sp_category('EX')
rl_sp_category('EX', parse = FALSE)
rl_sp_category_('EX')
}
}

