% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_sp_citation.R
\name{rl_sp_citation}
\alias{rl_sp_citation}
\alias{rl_sp_citation_}
\title{Get citations}
\usage{
rl_sp_citation(name = NULL, id = NULL, region = NULL, key = NULL,
  parse = TRUE, ...)

rl_sp_citation_(name = NULL, id = NULL, region = NULL, key = NULL, ...)
}
\arguments{
\item{name}{(character) Binomial taxonomic name}

\item{id}{(character) IUCN id}

\item{region}{(character) Binomial taxonomic name}

\item{key}{A IUCN API token}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\value{
A list, with the data in the \code{result} slot, unless using
a function with a trailing underscore, in which case json as character
string is returned.
}
\description{
Get citations
}
\examples{
\dontrun{
rl_sp_citation('Balaena mysticetus')
rl_sp_citation('Balaena mysticetus', region = 'europe')
rl_sp_citation(id = 12392)

rl_sp_citation(id = 2467, region = 'europe')
rl_sp_citation(id = 2467, region = 'europe', parse = FALSE)
rl_sp_citation_(id = 2467, region = 'europe')
}
}
\references{
API docs at \url{http://apiv3.iucnredlist.org/api/v3/docs}
}

