% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_sp_count.R
\name{rl_sp_count}
\alias{rl_sp_count}
\alias{rl_sp_count_}
\title{Get total species count of taxa in the Red List}
\usage{
rl_sp_count(key = NULL, parse = TRUE, ...)

rl_sp_count_(key = NULL, ...)
}
\arguments{
\item{key}{A IUCN API token}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
A list, with the data in the \code{result} slot, unless using
a function with a trailing underscore, in which case json as character
string is returned.
}
\description{
Get total species count of taxa in the Red List
}
\examples{
\dontrun{
rl_sp_count()
rl_sp_count_()
}
}
\references{
API docs at \url{http://apiv3.iucnredlist.org/api/v3/docs}
}

