% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_threats.R
\name{rl_threats}
\alias{rl_threats}
\alias{rl_threats_}
\title{Get species threats by taxon name, IUCN id, and region}
\usage{
rl_threats(name = NULL, id = NULL, region = NULL, key = NULL,
  parse = TRUE, ...)

rl_threats_(name = NULL, id = NULL, region = NULL, key = NULL, ...)
}
\arguments{
\item{name}{(character) Binomial taxonomic name}

\item{id}{(character) IUCN id}

\item{region}{(character) Binomial taxonomic name}

\item{key}{A IUCN API token}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
A list, with the data in the \code{result} slot, unless using
a function with a trailing underscore, in which case json as character
string is returned.
}
\description{
Get species threats by taxon name, IUCN id, and region
}
\examples{
\dontrun{
rl_threats('Fratercula arctica')
rl_threats('Fratercula arctica', region = 'europe')
rl_threats(id = 12392)
rl_threats(id = 22694927, region = 'europe')
rl_threats_('Fratercula arctica')

rl_threats(id = 62290750)
}
}
\references{
API docs at \url{http://apiv3.iucnredlist.org/api/v3/docs}
}

