% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_key.R
\name{rl_use_iucn}
\alias{rl_use_iucn}
\title{Helper to get and save IUCN API key}
\usage{
rl_use_iucn()
}
\description{
Browse IUCN Red List API key request URL and
provides instruction on how to store the key.
}
\details{
Note that after filling the online form, you should
receive an API key shortly but not immediately.
}
