% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_version.R
\name{rl_version}
\alias{rl_version}
\title{Get the Red List API version}
\usage{
rl_version(key = NULL, ...)
}
\arguments{
\item{key}{A IUCN API token}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\value{
API version as character string
}
\description{
Get the Red List API version
}
\examples{
\dontrun{
rl_version()
}
}

