% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_version.R
\name{rl_version}
\alias{rl_version}
\title{Get the Red List API version}
\usage{
rl_version(...)
}
\arguments{
\item{...}{Curl options passed to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
API version as character string
}
\description{
Get the Red List API version
}
\examples{
\dontrun{
rl_version()
}
}
